# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Painting Mistery API",
        "description": "Documentación de la API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/detalle-venta-producto": {
            "get": {
                "tags": [
                    "Detalle Venta Producto"
                ],
                "summary": "Listar detalles de productos vendidos",
                "operationId": "990404f7f4d4bb61391b73b60dd80ee7",
                "responses": {
                    "200": {
                        "description": "Lista"
                    }
                }
            },
            "post": {
                "tags": [
                    "Detalle Venta Producto"
                ],
                "summary": "Agregar producto a una venta",
                "operationId": "7c5380907d149195f3e619536236b891",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "venta_id",
                                    "producto_id",
                                    "cantidad"
                                ],
                                "properties": {
                                    "venta_id": {
                                        "type": "integer"
                                    },
                                    "producto_id": {
                                        "type": "integer"
                                    },
                                    "cantidad": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Agregado"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "400": {
                        "description": "Error de negocio"
                    }
                }
            }
        },
        "/detalle-venta-producto/{id}": {
            "get": {
                "tags": [
                    "Detalle Venta Producto"
                ],
                "summary": "Obtener detalle",
                "operationId": "7d7b12368d3c883833821173caf06cb3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Encontrado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Detalle Venta Producto"
                ],
                "summary": "Actualizar detalle",
                "operationId": "4106922f387c05f001548a7b2721d444",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "cantidad": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualizado"
                    },
                    "404": {
                        "description": "No encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Detalle Venta Producto"
                ],
                "summary": "Eliminar detalle",
                "operationId": "6351357d55b25f8d889817db405d54e5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Eliminado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            }
        },
...
    },
...
}